+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Saving thumbnails for faster startup
 - Start using message tab to display status information, etc.
 - Added support for more image/video formats (depending on QPixmap, vlc, cv2)
+- Task to print files that are in the default dirs, but not in the db
+  (to make cleanup easier)
 
 ## [0.3] - 2025-01-18